Initial commit
This commit introduces one post, one kb, and a minimal
about/verification page. This is mostly just to establish some
structure, and to publish something so that I don't end up working on
"Initial Commit" for so long that I lose motivation to upload something.

Non content files: a license file, a small tool to generate the
verification content, my public key for signature verification, hugo
config and layout info.